# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Toya Amerta API",
        "description": "REST API untuk aplikasi mobile (Flutter) Sistem Informasi PDAM Desa Toya Amerta. Semua response menggunakan envelope { success, message, data, meta }.",
        "contact": {
            "name": "Toya Amerta Dev Team",
            "email": "pt.andisuartika@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/petugas/dashboard": {
            "get": {
                "tags": [
                    "Petugas - Catat Meter"
                ],
                "summary": "Ringkasan dashboard petugas",
                "description": "Statistik pelanggan aktif, pencatatan meter, tagihan belum bayar, total kas terkumpul, serta daftar gabungan pelanggan sudah/belum dicatat untuk periode tertentu.",
                "operationId": "8139ae47f80e78fc8dffa7e427e4d209",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Tahun periode (default: tahun sekarang)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2026
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Bulan periode 1-12 (default: bulan sekarang)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 6
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "properties": {
                                                "active_customers": {
                                                    "type": "integer",
                                                    "example": 45
                                                },
                                                "recorded_count": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "not_recorded_count": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "unpaid_count": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "total_collected": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 540000
                                                },
                                                "history": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "water_reading_id": {
                                                                "type": "integer",
                                                                "example": 55,
                                                                "nullable": true
                                                            },
                                                            "customer_id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "customer_name": {
                                                                "type": "string",
                                                                "example": "Andi Suartika"
                                                            },
                                                            "customer_number": {
                                                                "type": "string",
                                                                "example": "PLG-2024-001"
                                                            },
                                                            "zone": {
                                                                "type": "string",
                                                                "example": "Lingkungan Sangket"
                                                            },
                                                            "recorded": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "reading_date": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2026-06-03",
                                                                "nullable": true
                                                            },
                                                            "total_amount": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 31000,
                                                                "nullable": true
                                                            },
                                                            "amount_paid": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 0,
                                                                "nullable": true
                                                            },
                                                            "payment_status": {
                                                                "type": "string",
                                                                "example": "belum_bayar",
                                                                "enum": [
                                                                    "belum_bayar",
                                                                    "sebagian",
                                                                    "lunas",
                                                                    "belum_dicatat"
                                                                ]
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "year": {
                                                    "type": "integer",
                                                    "example": 2026
                                                },
                                                "month": {
                                                    "type": "integer",
                                                    "example": 6
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Role tidak diizinkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthorizedResponse": {
                "title": "Unauthenticated Response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated. Token tidak valid atau sudah kedaluwarsa."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ForbiddenResponse": {
                "title": "Forbidden Response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Akses ditolak. Role Anda tidak memiliki izin."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Gunakan token yang didapat dari endpoint /auth/login. Format header: Authorization: Bearer {token}",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}